refactor(checker): split ProgramContext out of context/mod.rs
context/mod.rs had reached the 2000-line hard file-size limit (CLAUDE.md
§19), blocking any further CheckerContext field, doc comment, or impl
addition behind the arch-guard check.

Extract the cohesive, self-contained ProgramContext concern — the
project-wide shared state plus its apply_to/build_global_indices
orchestration — into a new sibling module context/program_context.rs and
re-export it via `pub use program_context::ProgramContext;`. The public
path tsz_checker::context::ProgramContext is unchanged.

This is a pure code-organization move with no behavior change. mod.rs
drops from 2000 to 1416 lines, leaving durable headroom for the next
CheckerContext field addition that motivated the issue.

Closes #11817
